+++ ARMeilleure/Decoders/OpCodeSimdHelper.cs
@@ -38,7 +38,7 @@
                     case 2:
                         // 2 x 32-bits floating point Immediate.
                         size = 3;
-                        imm = (long)DecoderHelper.Imm8ToFP32Table[(int)imm];
+                        imm = DecoderHelper.Imm8ToFP32Table[(int)imm];
                         imm |= imm << 32;
                         break;
