+++ ARMeilleure/Instructions/NativeInterface.cs
@@ -64,12 +64,12 @@ namespace ARMeilleure.Instructions
         #region "System registers"
         public static ulong GetCtrEl0()
         {
-            return (ulong)GetContext().CtrEl0;
+            return GetContext().CtrEl0;
         }
 
         public static ulong GetDczidEl0()
         {
-            return (ulong)GetContext().DczidEl0;
+            return GetContext().DczidEl0;
         }
 
         public static ulong GetTpidrEl0()
@@ -114,7 +114,7 @@ namespace ARMeilleure.Instructions
 
         public static void SetTpidrEl032(uint value)
         {
-            GetContext().TpidrEl0 = (long)value;
+            GetContext().TpidrEl0 = value;
         }
         #endregion
